Update main.py
I opened up your project and i realised there was nothing inside of the main.py file. instead contacting you i just realised i worked on a part of it about an hour ago so i decided why not place it in for you, if this looks good i can leave it on so we can work on it together. kinda cheap but something to me Lol